refactor(calendar): refactor calendar views and date helpers
Replaced MonthCalendarView with a unified CalendarView, moving date grid and formatting logic into a new Date+Ext.swift helper. Updated CalendarScreenView to use the new CalendarView and removed the obsolete Date+Formatting.swift and MonthCalendarView.swift files. Adjusted MainTabView to present CalendarScreenView instead of the old CalendarView.
